Update README and refactor undo support class
README updated to clarify BonsaiNumbering's features and integration with Blender and Bonsai, including improved descriptions of supported IFC data management. In numbering_tool.py, the UndoSupport class is renamed to UndoOperator, and all references are updated for consistency. Moreover, an update of the active object ensures that the number assignment is immediately visible in the Object Information tab.